+++ src/Aydsko.iRacingData/Common/DataResponse.cs
@@ -1,7 +1,7 @@
 ﻿// © 2022 Adrian Clark
 // This file is licensed to you under the MIT license.
 
-namespace Aydsko.iRacingData;
+namespace Aydsko.iRacingData.Common;
 
 public class DataResponse<TData>
 {
