+++ src/Aydsko.iRacingData/Common/LinkResult.cs
@@ -3,7 +3,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Aydsko.iRacingData;
+namespace Aydsko.iRacingData.Common;
 
 internal class LinkResult
 {
